aarch64: xfail pr109072_1.c's s16x4_2 [PR117092]
The fix for this depends on much more infrastructure which won't
be done for another few weeks. Pengxuan is working on the fix for GCC 16.
So let's xfail the testcase since it is a minor code quality regression.
we get:
```
        movi    v0.2s, 0
        ins     v0.h[0], w0
```
vs what we should get:
```
        and     x0, x0, 65535
        fmov    d0, x0
```
or
```
        fmov    h0, x0
```

Tested for aarch64-linux-gnu.

	PR target/117092

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/pr109072_1.c: xfail s16x4_2.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>