doc: Clarify REG_EH_REGION note usage
The documentation for the REG_EH_REGION could easily be read
(especially by non-native speakers) to indicate that it should be
attached to insn at the destination of an excpetion edge.  Despite the
original text saying that the note "specifies the destination," it is
actually always attached to the source instruction.

This updates the documentation to make it clear that the REG_EH_REGION
note is always attached to instructions originating an exception edge
and that the value of the note specifies where the exception edge
leads to.

Co-Developed-by: Philipp Tomsich <philipp.tomsich@vrull.eu>

gcc/ChangeLog:

	* doc/cfg.texi: Update the exception handling section for the
	REG_EH_REGION notes to make it clear that the note is attached
	to the instruction throwing the exception.