cobol: Remove dead code from the driver.
EXEC_LIB is no longer defined anywhere, so that the dependent code is
now dead.  The rpath code (if required) would need to be conditional
since rpath support is not available on all hosts.  Remove now and then
address if/when it is needed (other drivers do not do this).

gcc/cobol/ChangeLog:

	* gcobolspec.cc (append_rpath): Remove.
	(lang_specific_driver): Remove hard-wired rpath and library
	names.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>